[Loom] Extract bounded symbolic expressions
Move the allocation-free bounded expression summary into its own
analysis component. Source memory planning now depends on the narrow
bounded API instead of acquiring the arena-backed proof machinery
through the monolithic symbolic expression library.

Keep bounded normalization, traversal limits, and caller-owned storage
together with focused tests. A shared fixture preserves the production
IR and fact-table setup while allowing subsequent expression and proof
components to own their tests independently.
