+++ loom/src/loom/analysis/symbolic_expr_test.cc
@@ -12,6 +12,7 @@
 #include "iree/testing/gtest.h"
 #include "iree/testing/status_matchers.h"
 #include "loom/analysis/condition_facts.h"
+#include "loom/analysis/symbolic_expr_test_fixture.h"
 #include "loom/ir/attribute.h"
 #include "loom/ir/context.h"
 #include "loom/ir/module.h"
@@ -23,87 +24,6 @@
 
 namespace loom {
 namespace {
-
-class SymbolicExprTest : public ::testing::Test {
- protected:
-  void SetUp() override {
-    iree_arena_block_pool_initialize(4096, iree_allocator_system(),
-                                     &block_pool_);
-    iree_arena_initialize(&block_pool_, &analysis_arena_);
-
-    loom_context_initialize(iree_allocator_system(), &context_);
-    iree_host_size_t index_vtable_count = 0;
-    const loom_op_vtable_t* const* index_vtables =
-        loom_index_dialect_vtables(&index_vtable_count);
-    IREE_ASSERT_OK(loom_context_register_dialect(&context_, LOOM_DIALECT_INDEX,
-                                                 index_vtables,
-                                                 (uint16_t)index_vtable_count));
-    iree_host_size_t scalar_vtable_count = 0;
-    const loom_op_vtable_t* const* scalar_vtables =
-        loom_scalar_dialect_vtables(&scalar_vtable_count);
-    IREE_ASSERT_OK(loom_context_register_dialect(
-        &context_, LOOM_DIALECT_SCALAR, scalar_vtables,
-        (uint16_t)scalar_vtable_count));
-    iree_host_size_t scf_vtable_count = 0;
-    const loom_op_vtable_t* const* scf_vtables =
-        loom_scf_dialect_vtables(&scf_vtable_count);
-    IREE_ASSERT_OK(loom_context_register_dialect(
-        &context_, LOOM_DIALECT_SCF, scf_vtables, (uint16_t)scf_vtable_count));
-    IREE_ASSERT_OK(loom_context_finalize(&context_));
-
-    IREE_ASSERT_OK(loom_module_allocate(&context_, IREE_SV("test"),
-                                        &block_pool_, NULL,
-                                        iree_allocator_system(), &module_));
-    loom_builder_initialize(module_, &module_->arena,
-                            loom_module_block(module_), &builder_);
-    IREE_ASSERT_OK(
-        loom_value_fact_table_initialize(&fact_table_, &analysis_arena_, 16));
-    loom_symbolic_expr_context_initialize(
-        module_, &fact_table_, &analysis_arena_, &expression_context_);
-  }
-
-  void TearDown() override {
-    loom_module_free(module_);
-    loom_context_deinitialize(&context_);
-    iree_arena_deinitialize(&analysis_arena_);
-    iree_arena_block_pool_deinitialize(&block_pool_);
-  }
-
-  loom_value_id_t DefineIndexValue() {
-    loom_value_id_t value_id = LOOM_VALUE_ID_INVALID;
-    IREE_CHECK_OK(loom_builder_define_value(
-        &builder_, loom_type_scalar(LOOM_SCALAR_TYPE_INDEX), &value_id));
-    return value_id;
-  }
-
-  loom_value_id_t DefineI64Value() {
-    loom_value_id_t value_id = LOOM_VALUE_ID_INVALID;
-    IREE_CHECK_OK(loom_builder_define_value(
-        &builder_, loom_type_scalar(LOOM_SCALAR_TYPE_I64), &value_id));
-    return value_id;
-  }
-
-  void DefineFacts(loom_value_id_t value_id, loom_value_facts_t facts) {
-    IREE_CHECK_OK(loom_value_fact_table_define(&fact_table_, value_id, facts));
-    loom_symbolic_expr_context_reset(&expression_context_);
-  }
-
-  loom_op_t* BuildIndexConstant(int64_t value) {
-    loom_op_t* op = nullptr;
-    IREE_CHECK_OK(loom_index_constant_build(
-        &builder_, loom_attr_i64(value),
-        loom_type_scalar(LOOM_SCALAR_TYPE_INDEX), LOOM_LOCATION_UNKNOWN, &op));
-    return op;
-  }
-
-  iree_arena_block_pool_t block_pool_;
-  iree_arena_allocator_t analysis_arena_;
-  loom_context_t context_;
-  loom_module_t* module_ = nullptr;
-  loom_builder_t builder_;
-  loom_value_fact_table_t fact_table_;
-  loom_symbolic_expr_context_t expression_context_;
-};
 
 TEST_F(SymbolicExprTest, UnknownValueIsMemoizedLinearTerm) {
   loom_value_id_t value_id = DefineIndexValue();
@@ -530,97 +450,6 @@ TEST_F(SymbolicExprTest, ExpandsIndexMaddWithConstantMultiplier) {
   EXPECT_EQ(expression.terms[1].value_id, column);
 }
 
-TEST_F(SymbolicExprTest, BoundedExpressionPreservesOpaqueAssumeResult) {
-  loom_value_id_t row = DefineIndexValue();
-  loom_predicate_t predicate = {
-      /*.kind=*/LOOM_PREDICATE_RANGE,
-      /*.arg_count=*/3,
-      /*.arg_tags=*/
-      {LOOM_PRED_ARG_VALUE, LOOM_PRED_ARG_CONST, LOOM_PRED_ARG_CONST},
-      /*.reserved=*/{},
-      /*.args=*/{row, 0, 64},
-  };
-  loom_type_t index_type = loom_type_scalar(LOOM_SCALAR_TYPE_INDEX);
-  loom_op_t* assume_op = nullptr;
-  IREE_ASSERT_OK(loom_index_assume_build(&builder_, &row, 1, &predicate, 1,
-                                         &index_type, 1, LOOM_LOCATION_UNKNOWN,
-                                         &assume_op));
-  loom_value_id_t bounded_row = loom_index_assume_results(assume_op).values[0];
-
-  loom_symbolic_term_t terms[1] = {};
-  loom_symbolic_expr_t expression = {};
-  loom_symbolic_expr_from_value_bounded(module_, &fact_table_, bounded_row,
-                                        terms, IREE_ARRAYSIZE(terms),
-                                        &expression);
-
-  ASSERT_TRUE(loom_symbolic_expr_is_linear(&expression));
-  EXPECT_EQ(expression.constant, 0);
-  ASSERT_EQ(expression.term_count, 1);
-  EXPECT_EQ(expression.terms[0].coefficient, 1);
-  EXPECT_EQ(expression.terms[0].value_id, bounded_row);
-}
-
-TEST_F(SymbolicExprTest, BoundedExpressionExpandsAssumedAffineSource) {
-  loom_value_id_t row = DefineIndexValue();
-  loom_value_id_t four = loom_index_constant_result(BuildIndexConstant(4));
-  loom_op_t* shifted_op = nullptr;
-  loom_type_t index_type = loom_type_scalar(LOOM_SCALAR_TYPE_INDEX);
-  IREE_ASSERT_OK(loom_index_add_build(&builder_, row, four, index_type,
-                                      LOOM_LOCATION_UNKNOWN, &shifted_op));
-  loom_value_id_t shifted = loom_index_add_result(shifted_op);
-  loom_predicate_t predicate = {
-      /*.kind=*/LOOM_PREDICATE_RANGE,
-      /*.arg_count=*/3,
-      /*.arg_tags=*/
-      {LOOM_PRED_ARG_VALUE, LOOM_PRED_ARG_CONST, LOOM_PRED_ARG_CONST},
-      /*.reserved=*/{},
-      /*.args=*/{shifted, 0, 64},
-  };
-  loom_op_t* assume_op = nullptr;
-  IREE_ASSERT_OK(loom_index_assume_build(&builder_, &shifted, 1, &predicate, 1,
-                                         &index_type, 1, LOOM_LOCATION_UNKNOWN,
-                                         &assume_op));
-  loom_value_id_t bounded_shifted =
-      loom_index_assume_results(assume_op).values[0];
-
-  loom_symbolic_term_t terms[1] = {};
-  loom_symbolic_expr_t expression = {};
-  loom_symbolic_expr_from_value_bounded(module_, &fact_table_, bounded_shifted,
-                                        terms, IREE_ARRAYSIZE(terms),
-                                        &expression);
-
-  ASSERT_TRUE(loom_symbolic_expr_is_linear(&expression));
-  EXPECT_EQ(expression.constant, 4);
-  ASSERT_EQ(expression.term_count, 1);
-  EXPECT_EQ(expression.terms[0].coefficient, 1);
-  EXPECT_EQ(expression.terms[0].value_id, row);
-}
-
-TEST_F(SymbolicExprTest, BoundedExpressionNormalizesCancellation) {
-  loom_value_id_t row = DefineIndexValue();
-  loom_value_id_t column = DefineIndexValue();
-  loom_type_t index_type = loom_type_scalar(LOOM_SCALAR_TYPE_INDEX);
-  loom_op_t* sum_op = nullptr;
-  IREE_ASSERT_OK(loom_index_add_build(&builder_, row, column, index_type,
-                                      LOOM_LOCATION_UNKNOWN, &sum_op));
-  loom_op_t* difference_op = nullptr;
-  IREE_ASSERT_OK(loom_index_sub_build(&builder_, loom_index_add_result(sum_op),
-                                      column, index_type, LOOM_LOCATION_UNKNOWN,
-                                      &difference_op));
-
-  loom_symbolic_term_t terms[1] = {};
-  loom_symbolic_expr_t expression = {};
-  loom_symbolic_expr_from_value_bounded(
-      module_, &fact_table_, loom_index_sub_result(difference_op), terms,
-      IREE_ARRAYSIZE(terms), &expression);
-
-  ASSERT_TRUE(loom_symbolic_expr_is_linear(&expression));
-  EXPECT_EQ(expression.constant, 0);
-  ASSERT_EQ(expression.term_count, 1);
-  EXPECT_EQ(expression.terms[0].coefficient, 1);
-  EXPECT_EQ(expression.terms[0].value_id, row);
-}
-
 TEST_F(SymbolicExprTest, DynamicMultiplyFallsBackToResultSymbol) {
   loom_value_id_t left = DefineIndexValue();
   loom_value_id_t right = DefineIndexValue();
@@ -750,28 +579,6 @@ TEST_F(SymbolicExprTest, ProvesFlattenedAddressFromDynamicAxisBounds) {
   IREE_ASSERT_OK(loom_symbolic_expr_prove_le(
       &expression_context_, &exclusive_end, &element_count, &proof));
   EXPECT_EQ(proof, LOOM_SYMBOLIC_PROOF_UNKNOWN);
-}
-
-TEST_F(SymbolicExprTest, BoundedDynamicMaddFallsBackToResultSymbol) {
-  loom_value_id_t left = DefineIndexValue();
-  loom_value_id_t right = DefineIndexValue();
-  loom_value_id_t addend = DefineIndexValue();
-  loom_op_t* madd_op = nullptr;
-  IREE_ASSERT_OK(loom_index_madd_build(&builder_, left, right, addend,
-                                       loom_type_scalar(LOOM_SCALAR_TYPE_INDEX),
-                                       LOOM_LOCATION_UNKNOWN, &madd_op));
-  loom_value_id_t result = loom_index_madd_result(madd_op);
-
-  loom_symbolic_term_t terms[2] = {};
-  loom_symbolic_expr_t expression = {};
-  loom_symbolic_expr_from_value_bounded(module_, &fact_table_, result, terms,
-                                        IREE_ARRAYSIZE(terms), &expression);
-
-  ASSERT_TRUE(loom_symbolic_expr_is_linear(&expression));
-  EXPECT_EQ(expression.constant, 0);
-  ASSERT_EQ(expression.term_count, 1);
-  EXPECT_EQ(expression.terms[0].coefficient, 1);
-  EXPECT_EQ(expression.terms[0].value_id, result);
 }
 
 TEST_F(SymbolicExprTest, ProvesLessEqualFromTermFacts) {
